Moving everything into base.py
